requirements.txt
streamlit
pandas
numpy
plotly
ytmusicapi
scikit-learn
transformers
torch
tqdm
joblibq